[ocean] Fix error checking the use of 'from_date' param
The code used to check whether a backend uses or not
'from_date' parameter was invalid. It checked if the
backend had that attribute but so far, none of Perceval
backends store it as a instance attribute.

The current method uses introspection, checking the
signature of the 'fetch()' method to determinate the use
of 'from_date'.