Changes grimoire to grimoire_elk.
Now, the name of the module is grimoire_elk. The Python pacakge
is grimoire-elk. The util/p2o.py script, and other scripts in util
have been changed to reflect this.

Any script or library using grimoire (as it was called) should
now change to import grimoire_elk.

The util/p2o.py script is installed as an executable scritp.