Add dependency on perceval-mozilla
perceval-mozilla is needed fpr p2o to run. For now, I'm including
it as a dependency. Probably we should make grimoire_elk a bit
smarter, and use perceval-mozilla only if it is available, or
else, throw an exception if it is invoked when calling the module,
but it is not installed.

I bumped version to 0.22.1 to upload a new package to pypi.