[filtering] Use origin per repository filtering except for stackexchange where tag is used
Until all data sources have tag info in raw and enrich indexes the filtering using
the tag field is only included for stackexchange because this won't work without it.
The filter creation is now in a common method get_repository_filter so it will be easy
to migrate to tag based filtering.