[enrich][phabricator] Undo eitem['status'] change to use the upercase value for status Open and the rest.
In the original specification for the enrichment of maniphest the status
value used was phab_item['fields']['status']['name']. It was changed later
to phab_item['fields']['status']['value'] to use the same status value that
appears in other places, but this has broken the maniphest panel.
So undoing this change.