[kidash] Add support for using panels in the panels module
Now, panels (JSON files with panel defintions) are looked for
first in a local directory panels/json, and if not found, in the
panels module, if it is present and importable.
For maintaining compatibility, panels with names such as
"panels/json/git.json" are supported. But the most simple
"git.json" is also supported, meaning "git.json" file in the
current directory, or "git.json" file in the corresponding
directory of the modules file.

See README.md for more details.

Versions for both grimoire-elk and grimoire-kidash are updated.

grimoirelab-panels is now a dependency of grimoire-kidash.