[enrich] Use _version.py to get GELK version
Version was obtained by getting the git commmit description,
using the git command. Instead of that, use the _version.py file
that we maintain for consistent versioning.

The previous schema, based on the git command, does not work
with pip packages, since they are not clones of the git repo.