[utils] Add index_mapping.py script
This script copy one GrimoireLab index to other index ensuring that
the mappings are right in the process. It is specially useful in migrations
to create the right mapping for an index.

[utils lib] Add new method get_connector_name_from_cls_name

It returns the name of a connector given the name of a class that
is part of it.