[panels] Add support for listing panels in ES6
In ES6 the types in the indexes has been removed. And .kibana
does not have anymore the types dashboard, visualization, search
and index-pattern. The types are now included inside the items.
The only type in .kibana is now doc.

This commit add the code to detect the version of ES and adjust
howto work the .kibana index in ES6.