Merge remote-tracking branch 'upstream/elastic6' into elastic6
Resolved merging.
Added a new function in ElasticSearch class
to detect the release of the Elasticsearch instance.
Added testing, for that new class.
Added check when uploading default mapping, depending
on the version of Elasticsearch.
Changed "string" to "keyword" or "text" (depending if analyzed
or not analyzed) in mappings for backends that needed it.
Added some debug messages for better tracking when uploading
and fetching items.