[elasticsearch] Support Elasticsearch 6
Add changes to support Elasticsearch 6, while maintaining compatibility
with Elasticsearch 5 and 2. In fact, compatibility with Elasticsearch 2
is now improved, supporting not only raw indexes, but also enriched indexes.

New Mapping class, for mappings. Solves the problem of neededing to know
the version of Elasticsearch before uploading the mapping.

Minor changes to how test.conf is read, so that now you can include
a section for a database (user, passwd).

Fix for access to twitter raw indexes.

In my laptop tests are passed with ES 6.0, 5.0, and 2.4.