[tests] Add tests using csv mappings
This code includes tests to check that the mappings defined
in the enriched indexes are aligned with the ones used in
the corresponding panels (provided as csv files available in
the schema folder). The tests  are executed in the method
`_test_raw_to_enrich` when sortinghat is used. It leverages
on the schema and comparator provided by
grimoirelab-sigils/owlwatch.