[studies] Add enrich_pull_requests study
This code adds the `enrich_pull_requests` study to the enriched/github.py
file. This study aims at adding additional fields that are related to
pull request items but are not fetched by perceval directly for items of the
pull_requests category.
This study should be used using the functionality added by the PR:
chaoss/grimoirelab-sirmordred#190.

This study takes in an input param: `raw_issues_index` which should contain
the data fetched from the same repository from which the pull requests data
is fetched. Using this `raw_issues_index`, the additional data is extracted from
related issue and that data is enriched and added to the relevant pull request
and updated in the enriched pull request index.

This code also:
- uses `str_to_datetime` function to convert date in string format to datetime
where ever time difference between two dates is being calculated
- corrects typo in enriched/git.py