[enriched] Update mediawiki project json params
This change allows to add a second URL to `projects.json`
file for mediawiki.

In that case, the first URL will be used for retrieving raw data
using Perceval, while the second one will be used to build
the URL of the wiki page. If both, API and web server, uses
the same URL, it is not needed to write the same URL twice,
just let one URL as parameter instead of two.