[onion] Add refresh interval
A new parameter `seconds` can be used to define the minimum
time from one study execution to the following. By default has
been set to one week, but can be set for each data source in
`setup.cfg`. Onion will only run again if at least that number
of seconds has passed from the enrichment of the most recent
item in the index.

For information purposes, each time the study is not run due
to this parameter, a log line is printed containing that
minimum date for reference.

Note that the study won't be executed exactly at that time,
but at some point after that time. Studies are executed
after each enrichment processes.