[onion] Fix `all_onion` alias creation
Alias must be created if it doesn't exist for the
corresponding `out_index`. There was a bug when checking it
because code was looking only for alias to exist. As
`all_onion` points to different indices (one per data source)
it could exist pointing to another source, which was a false
positive that prevented the alias to be created.

This commit allows to check if the alias exists pointing to
the particular index we are interested in.