[raw-jira] Select jira project from projects.json url
This code allows to read the project param of the Jira
Perceval backend from the corresponding projects.json url. Thus,
it enables to collect and enrich only issues of specific projects
within the Jira server.

In order to use this feature, the url within the projects.json
should be defined in the following way (ODP is the name of the
Jira project):
```"jira": [
           "http://finosfoundation.atlassian.net ODP"
        ], ..
```