[git] Remove deleted commits from raw and enriched indexes
This code retrieves the commits not present in the original
repository and delete the corresponding documents from the raw and
enriched indexes. In a nutshell, at the end of the enrichement phase,
 the list of current commit hashes are retrieved from the repository
and compared with the hashes stored in the raw index. The different of
these two sets gives the deleted commits (the ones that are stored
but not present in the repo), which are removed from the
raw and enriched indexes.