[elastic_items] Update ES query for incremental enrichment
This code changes the behaviour of the ES query to retrive
items from the raw indexes for incremental enrichment purposes.
In a nutshell, the query now uses a QUERY context which returns
the items matching the query plus a score that rates how good
the match is. However the score is not needed in ELK, thus the
content is changed to FILTER context, which doesn't calculate
any score.