[enrich] Update Git references data in raw and enriched indexes
This code updates the information about Git references
of the items corresponding to commits in both raw and enriched
indexes. The update is needed due to operations (e.g., merges
and branch deletions) which affect the repository references
and consequently their commits. Thus, this code aims at
ensuring consistency between the information of the original
Git repository and the data stored in ElasticSearch.