[enrich] Update branch info in git commits
This code allows to update the branch information of the JSON documents representing
Git commits. First the list of refs is cleared in the enriched index, then
the original repo is queried to retrieve the list of commit hashes for each branch and
finally the obtained hashes are added to the refs attribute of the documents.