[raw] Parse filters passed in the projects.json entries
This code allows to parse the filters declared in the
projects.json entries. This code applies for the Perceval
backends that require only an URL as input parameter, for
the others this method should be redefined.

Filters are identified by the delimiter `--filter`, while the filter
names and values are separated by `=`. For instance:
`--filter-raw=data.product:Add-on SDK,data.component:General`