[elk] Store identities in use in dedicated index
This code populates the index `grimoirelab-identities-cache`
with information about the unique identities seen in the
enriched indexes. Every unique identity identifier is stored
in the index together with the time when it was last seen.
For every enrich iteration, the last seen time is refreshed
in case the unique identity appears again.