[elk] Retain identities based on identities index data
This code allows to delete unique identities in SortingHat
based on the information contained in the grimoirelab-identities-cache
index. Given the retention time param, the date before which
an identity should be removed is calculated and used to
retrieve the list of unique identifiers contained in the identities
index. The obtained identitfiers are finally used to call
the sortinghat command to delete unique identities.