[enrich-meetup] Avoid failures when 'rsvps' field is not available
Due 'rsvps' is a classified field, it might not be available
when items are retrieved. So far, this field is considered
as mandatory by the enricher.

With this patch, ELK won't break when this field is not found.

This patch also updates 'populate_identities_index' function
to ignore those items that don't have identities. That
happens in the case of Meetup, when 'event_host' or 'rsvps'
are removed and no other identities are available.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>