[enriched-jira] Add comments to enriched index
This code includes Jira comments to the enriched index. To do
this, changes have been made to allow the cohesistance of both
comments and issues in the same index, following an approach
similar to meetup and stackexchange enrichers. Thus, enriched
issues and commits have different ids, the former is the issue
ids, while the latter is a str concatenating the issue id and
the comment id.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>