[enriched-enrich] Refactor `find_item_project` method
This code refactors the logic of the method `find_item_project`,
which now relies on the variable `projects_json_repo` (containing
the repo url declared in the projects.json currently processed).
The project is identified by using the data source name plus the
value of `projects_json_repo`. In case the projects_json_repo is
None, the previous logic is used.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>