[elk] Set value of `projects_json_repo` during collection
This code sets the value of projects_json_repo during the
collection phase (`feed_backend` function). This change
allows to know which repo url (declared in the projects.json)
is being currently analyzed, thus it simplifies the mapping
between the project declared in the projects.json and the
repo included in it.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>