[enriched-utils] Add filter raw text to retrive last enrich date
This code includes the text representation (taken from the projects.json)
of the filter raw to the filters used to retrieve the last enrich date.
In a nutshell, this new filter allows to identify the last enrich date
of the items relying on the filter-raw param which were missing before a
clear correspondence with the raw items, thus making impossible to
retrieve the correct last enrich date in case of failures during the
enrichment process.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>