[enriched-gerrit] Add patchset approvals as enriched items
This code includes the patchset approvals to the enriched index. After
creating the enriched patchset, the approvals in the raw items are
processed and linked to the enriched review and enriched patchset by adding
the url, summary, repository, branch and number of the review, and the number,
revision of and ref of the enriched patchset.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>