[enriched-git] Add study to process git branches
This code defines a study to process git branches. It
consists in updating the branch information for each
doc (which represents a git commit) in the enriched index.

The study accesses the projects.json, retrieves the urls of
git repos and iterates them one by one to update branch info
within the documents.

The study is activated by adding the section [enrich_git_branches]
in the setup.cfg. This section doesn't require any parameter.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>