[enriched-enrich] Handle errors in demography study
This code allows to handle errors that may pop up
during the execution of the demography study. The
update_by_query underlying the study is executed
with the params `wait_for_completion=true` and
`conflicts=proceed`. The former waits that the
updates have finished, while the latter prevents
to stop the study in case of a version conflicts
in the documents updated. Furthermore, the request
issuing the update_by_query is wrapped in a
try/except block to ignore Retry errors, which
may occur when ElasticSearch is not available.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>