[elk] Improve log messages
This code improves the log messages of elk, in particular
at the end of a collection/enrichment step, the message
has been changed to give more details about the repo
just processed. Furthermore, some messages have been
rewritten to remove the term `ocean`, which relates
to a very intial implementation of elk.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>