[enriched-enrich] Fix conversion projects.json to map
This code fixes the way the projects map is built by
processing the repositories under the `unknown`
section at the very end. This change is needed to
avoid assigning repositories to the generic
project `Main` when they are also declared in other
projects.

This error appears with Python 3.5, but not with
Python 3.6 and it is related to how the dictionary
keys are ordered.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>