[elastic] Change logic to handle errors when creating a mapping
This code changes the logic of the method `create_mappings` when
handling errors during the mapping creation. Now the HTTP error
is intercepted and a log message is written down.

This change is needed to improve the coherence of the code, by
applying the same approach (i.e., try except) for similar cases.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>