[enriched-enrich] Add enrich_feelings study
This code includes the enrich_feelings study, which
allows to add sentiment and emotion data to a target
enriched index.

The study requires the URL of the NLP REST service,
plus a list of attributes containing the text to
be processed. The logic of the study is described below.

All items in the target enriched index not including
the fields `has_sentiment` or `has_emotion` are retrieved.
For each item, the target attribute is identified, then the
text is convert to plain text using the NLP REST endpoint
`plainTextBugTrackerMarkdown`. The obtained text is processed
to strip out code parts using the endpoint `code`, the text
left is passed to the endpoints `sentiment` and `emotion`,
which are used to derive sentiment and emotion labels for
the input text. The labels collected are saved in the
corresponding item on the attributes `feeling_sentiment` and
`feeling_emotion`. Furthermore, also the attributes
`has_sentiment = 1` and `has_emotion = 1` are added to the
item to prevent the recalculation of feeling study for
items already analysed.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>