[enriched-enrich] Add survival developer study
This code includes a new study, which aims at forecasting
the contributor activity based on their past contributions.
The idea behind this study is that abandonment of active
developers poses a significant risk for open source software
projects, and this risk can be reduced by forecasting the
future activity of contributors involved in such projects
and taking necessary countermeasures. The logic of this study
is based on the tool: https://github.com/AlexandreDecan/gap.

The study is executed for each repository in the enriched
index on tunable monthly timeframes. For each timeframe, the
authors are retrieved and the corresponding activity dates
collected. Such dates together with the param `probability`
are used to perform a survival analysis. The result of
the analysis is a list of survived developers that should
contribute in the next timeframe.

The study provides the parameters `observations` and
`max_recency` to filter authors with less than X activities
per timeframe and author which last activity happened more
than Y days before the end of the timeframe.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>