Add additional index to compute the chronological evolution of opened issues and average opened time issues.
For each repository and label, we start the study on repository
creation date until today with a day interval (default). For each date
we retrieve the number of open issue at this date by difference between
number of opened issues and number of closed issues. In addition, we
compute the average opened time for all issues open at this date.

To differenciate by label, we compute evolution for bugs and all others
labels (like "enhancement","good first issue" ... ), we call this
"reduced labels". We need to use theses reduced labels because the
complixity to compute evolution for each combinaison of labels would be
too big. In addition, we rename "bug" label to "bugs" in reduced labels.

Signed-off-by: Florent Kaisser <contact@lille.bike>