[grimoire_elk] Add support for multiple affiliations
This code allows to store in the enriched indexes information
of multiple affiliations. Such a data is stored:
- as a list in the sortinghat-related attribute `<rol>_multi_org_names`
  Ex.: author_multi_org_names
- as single sortinghat-related attributes `<rol>_multi_org_name_<num>`
  Ex.: author_multi_org_name_0, author_mult_org_name_1

The code that deals with the identities refresh has been modified as well
to make sure that the affiliations stored in the list `<rol>_multi_org_names`
is aligned with the attributes `<rol>_multi_org_name_<num>`. The change
consists of removing the attributes `<rol>_multi_org_name_<num>` and add
them again. It was need to cover the cases when the number of affiliations
decreases. In this scenario, the attributes `<rol>_multi_org_name_<num>`
with `<num`> greater than the new list size `<rol>_multi_org_names` were
not modified/updated, thus introducing a misalignment between the list
and the single attributes.

Tests have been updated accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>