[enriched-study-ceres-onion] Support for multiple affiliations
This code allows to create the onion study index using the
information contained in the `author_multi_org_names` attribute.

In a nutshell, the organizations are first retrieved from this
attribute. If no data is returned, the organizations are
retrieved from the attribute `author_org_name`. The latter
works like a fallback option (to ensure compatibility with
old enriched indexes not including `author_multi_org_names`).

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>