[github] Make owners data anonymization optional for github
This code anonymizes the information coming from Perceval related
with users. For each identity it only keeps a hash of the name and
login. Email, company, location and other information related with
the user is removed.
Important: Because user information is not required, it is important
to use the `filter-classfied` flag to avoid getting information from
the GitHub's User API
```
[github]
...
filter-classfied = true
anonymize = true
```

Signed-off-by: JJMerchante <jj.merchante@gmail.com>