[enriched-gerrit] Propagate changeset status to patchsets
This code copies the attributes `status` and `status_value` of
the changeset to the corresponding patchsets. With these new
attributes, it is possible to compare the time to first review
for patchsets belonging to a given type of changesets (e.g.,
merged, abandoned or new).

These new attributes are also included in the changeset
enriched items for consistency purpose.

Tests have been updated accordingly.
Schema has been updated accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>