[elk] Handle private repos in git studies
This code prevents to store the credentials of the
git private repos in the studies indexes
(aoc and git branches).

This change requires to improve also the way the
project attribute is assigned to a repository. In the
case of a private repo, the corresponding data source
in the projects.json is accessed, each repo is anonymized
and the output is compared with current repo (which value
is stored in the indexes).

Tests have been updated accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>