[enriched/githubql] Add new study 'enrich_reference_analysis'
The purpose of this study is to gather all the issues and
pull requests which are mutually referenced. Once these
references are obtained, all of the events for the given
issue or pull request are updated with the corresponding
list of URLs from the referenced items.

The following fields are added:
* `referenced_by_issues`: List of issues referenced by a
    given Issue or Pull Request, from the same repository.
* `referenced_by_prs`: List of pull requests referenced by
    a given Issue or Pull Request, from the same repository.
* `referenced_by_external_issues`: List of issues referenced
    by a given Issue or Pull Request, from different
    (external) repositories.
* `referenced_by_external_prs`: List of pull requests
    referenced by a given Issue or Pull Request, from
    different (external) repositories.

Unit tests added accordingly.

Schema updated accordingly.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>