[gerrit] Add last-verified approval value field to enricher
Before this change, the last approval value for a given changeset
was computed only for approvals labelled as "Code-Review". Now
this last approval value is calculated for approvals labelled as
"Verified".

The field is `changeset_status_value_verified` and it is defined
as the value of the last "Verified" approval made by someone else
than the author for a given changeset, it can be `-2`, `-1`, `1`
or `2`.

The CSV file containing the Gerrit schema has been updated
accordingly.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>