[gitlab] Autodetect gitlab enterprise-url arg
The `enterprise-url` parameter of the `gitlab` backend is currently left
to be specified by the user in the config file.  This commit removes
this burden from the user by automatically detecting the enterprise URL
from the url passed to get_perceval_params_from_url().

This aims to make it easier for users to work with enterprise GitLab
instances.  The `enterprise-url` parameter isn't very well documented
for a user who hasn't used perceval directly before, or to any users who
don't understand that the config can be used to pass arbitrary command
line args to perceval (the parameter isn't documented in the default
config or the list of supported backends/their parameters).

However, by detecting the parameter automatically, this becomes
something that the end user doesn't even need to worry about!

As far as I know, this does not introduce any backwards-compatibility
issues.

Ammended from fbb340f:  Update variable name -> host, add comment

Signed-off-by: Emi Simpson <emi@alchemi.dev>