[mattermost] Allow parsing normal mattermost urls
With the addition of
https://github.com/chaoss/grimoirelab-perceval/pull/748
to grimoirelab-perceval, it is now possible to call perceval using
mattermost team & channel names.  This means that it is possible to
accept a standard mattermost URL in grimoirelab-elk.  This PR adds
support for this ability, while maintaining support for "legacy" repos
in the old format.

This also adds a degree of validation to URLs, since they are now
matched by regex, and improperly formatted URLs will fail to match.  I
think this could be an opportunity to add an informative error message
linking to the docs, but I wasn't able to find a place where the URL
format is documented to link to.

Ammended from b5edb50:  Normalize capital letters before processing URLs
Ammended from 77b18dc:  Add sign-off
Ammended from 436f3ef:  Filled in TODO in error message

Signed-off-by: Emi Simpson <emi@alchemi.dev>