[enriched/github2] Use user_data fields to set identities values
Name and username fields on the identities were using
incorrect values. In the case of 'name' it was using 'login'
value and not the name under 'user_data'. For the username
it was using a field that doesn't not exist.

This patch fixes the problem using 'name' and 'login'
for the two fields.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>