[enriched/git] Add commit meta fields
This code extracts all meta info authors in the commit using
the new method `__add_commit_meta_fields` also creates a new
identity in SH if it does not exist:
- Acked-by
- Co-Developed-by
- Reported-by
- Reviewed-by
- Signed-off-by
- Suggested-by
- Tested-by

The new fields will be named the same as the meta info but in
lower case and for each of them these suffixes will be added:
- multi_bots
- multi_domains
- multi_names
- multi_org_names
- multi_uuids

Example for `Signed-off-by`:
- signed_off_by_multi_bots
- signed_off_by_multi_domains
- signed_off_by_multi_names
- signed_off_by_multi_org_names
- signed_off_by_multi_uuids

Also add the new method `get_item_sh_meta_fields` to get the
meta fields needed to refresh them automatically in the method
`refresh_identities`.

Test updated accordingly.
Schema updated accordingly.

Signed-off-by: Quan Zhou <quan@bitergia.com>