[SortingHat] Reduce the number of queries to SortingHat
This commit updates the code to remove the number of queries to
SortingHat and implements the lru cache. It improves the time
needed to enrich each item.

Given a sh_id or an identity, it collects all the information
in the same query, including the identity, profile, is_bot and all
the enrollments.

Update the tests accordingly.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>