Improve performance refreshing identities
This commit improves the process of refreshing identities.

The primary objective is to reduce the number of queries
made to SortingHat. The approach involves retrieving
individuals along with their identities and profile
information, which is then passed on to the refreshing
process, that way that information is not retrieved again.

This commit also improves the load identities process.
It skips the query for adding recently added identities.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>