[enrich] Fix enrich items with multiple enrollments
This commit fixes the date format when there are several
enrollments in the profile. Since the number of times the
`str_to_datetime` method is executed has been reduced to
improve performance, the `date` field in `get_item_sh_from_id`
and `get_item_sh_meta_fields` must be strings.

Test to enrich an item with multiple enrollments added.

Signed-off-by: Quan Zhou <quan@bitergia.com>