[kitsune] Include unique field in enriched index
This commit updates the Kitsune enriched index. It was not
including a unique identifier in the enriched index for questions
and answers, this was causing an error when trying to autorefresh.

Include also the default raw fields in answers items that were missing:
"metadata__updated_on", "metadata__timestamp","offset",
"origin", "tag", and "uuid".

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>