[git] Improve branches study for consistency
Previously, while the branches study was running,
the `branches` field remained empty or partially filled
until the study completed, leading to incorrect data
being displayed on the dashboard.
With this change, the study is first created in an
auxiliary field `branches_aux`, and once complete,
the results are moved to the `branches` field.
This approach ensures consistent and accurate data
is displayed.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>