[git aoc/onion] Implement custom bulk write for OpenSearch
The current bulk helper from OpenSearch tries to
decode the documents as UTF-8, and in many cases, there
are items that are not compatible and raise an exception.

Add a new bulk_write method for Cereslib `ESConnector` to handle
document uploads to ElasticSearch using the bulk API. This new method
manages bulk sizes and handles errors during the upload process.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>